投稿のプレビュー機能

+++ src/client/app/common/scripts/post-form.ts
@@ -45,6 +45,7 @@ export default (opts) => ({
 	data() {
 		return {
 			posting: false,
+			preview: null,
 			text: '',
 			files: [],
 			uploadings: [],
@@ -457,10 +458,11 @@ export default (opts) => ({
 			this.text += getFace();
 		},
 
-		post() {
+		post(preview: boolean) {
 			this.posting = true;
 			const viaMobile = opts.mobile && !this.$store.state.settings.disableViaMobile;
 			this.$root.api('notes/create', {
+				preview,
 				text: this.text == '' ? undefined : this.text,
 				fileIds: this.files.length > 0 ? this.files.map(f => f.id) : undefined,
 				replyId: this.reply ? this.reply.id : undefined,
@@ -480,6 +482,12 @@ export default (opts) => ({
 				} : null,
 				viaMobile: viaMobile
 			}).then(data => {
+				if (preview) {
+					this.preview = data.createdNote;
+					this.preview.id = Math.random();
+					return;
+				}
+				this.preview = null;
 				this.clear();
 				this.deleteDraft();
 				this.$emit('posted');
